+++ lib/src/element/sub/button/radio_btn_element.dart
@@ -49,6 +49,9 @@ class RadioBtnElement extends MultiChildElement {
   @override
   RadioBtnElement initParams() {
     super.initParams();
+    elParams.p.isEnabled = params.containsKey(EnumSpWMLParams.isEnabled)
+        ? params[EnumSpWMLParams.isEnabled]
+        : true;
     SelectableIconBtnParams enabled = SelectableIconBtnParams();
     enabled.icon = const Icon(Icons.radio_button_checked_outlined);
     enabled.iconSize = params.containsKey(EnumSpWMLParams.iconSize)
@@ -126,6 +129,15 @@ class RadioBtnElement extends MultiChildElement {
   void setManager(IndexManager m) {
     elParams.p.manager = m;
   }
+
+  /// (en) Enable/disable this button.
+  ///
+  /// (ja)このボタンの有効・無効を切り替えます。
+  ///
+  /// * [isEnabled] : If true, the button is enabled.
+  void setEnabled(bool isEnabled) {
+    elParams.p.isEnabled = isEnabled;
+  }
 }
 
 class _RadioBtnElementWidget extends StatefulWidget {
@@ -157,9 +169,11 @@ class _RadioBtnElementWidgetState extends State<_RadioBtnElementWidget> {
     if (widget.elParams.p.enableTapLabel) {
       return InkWell(
         key: widget.elParams.p.enableTapInkWellParams.key,
-        onTap: () {
-          _onTapCallback(index);
-        },
+        onTap: widget.elParams.p.isEnabled
+            ? () {
+                _onTapCallback(index);
+              }
+            : null,
         onDoubleTap: widget.elParams.p.enableTapInkWellParams.onDoubleTap,
         onLongPress: widget.elParams.p.enableTapInkWellParams.onLongPress,
         onTapDown: widget.elParams.p.enableTapInkWellParams.onTapDown,
@@ -208,14 +222,12 @@ class _RadioBtnElementWidgetState extends State<_RadioBtnElementWidget> {
         hoverColor: widget.elParams.p.enableParams!.hoverColor,
         highlightColor: widget.elParams.p.enableParams!.highlightColor,
         splashColor: widget.elParams.p.enableParams!.splashColor,
-        disabledColor: widget.elParams.p.enableTapLabel
-            ? widget.elParams.p.enableParams!.color
-            : widget.elParams.p.enableParams!.disabledColor,
-        onPressed: widget.elParams.p.enableTapLabel
-            ? null
-            : () {
+        disabledColor: widget.elParams.p.enableParams!.disabledColor,
+        onPressed: widget.elParams.p.isEnabled
+            ? () {
                 _onTapCallback(index);
-              },
+              }
+            : null,
         mouseCursor: widget.elParams.p.enableParams!.mouseCursor,
         focusNode: widget.elParams.p.enableParams!.focusNode,
         autofocus: widget.elParams.p.enableParams!.autofocus,
@@ -240,14 +252,12 @@ class _RadioBtnElementWidgetState extends State<_RadioBtnElementWidget> {
         hoverColor: widget.elParams.p.enableParams!.hoverColor,
         highlightColor: widget.elParams.p.enableParams!.highlightColor,
         splashColor: widget.elParams.p.enableParams!.splashColor,
-        disabledColor: widget.elParams.p.enableTapLabel
-            ? widget.elParams.p.enableParams!.color ?? Colors.black87
-            : widget.elParams.p.enableParams!.disabledColor,
-        onPressed: widget.elParams.p.enableTapLabel
-            ? null
-            : () {
+        disabledColor: widget.elParams.p.enableParams!.disabledColor,
+        onPressed: widget.elParams.p.isEnabled
+            ? () {
                 _onTapCallback(index);
-              },
+              }
+            : null,
         mouseCursor: widget.elParams.p.enableParams!.mouseCursor,
         focusNode: widget.elParams.p.enableParams!.focusNode,
         autofocus: widget.elParams.p.enableParams!.autofocus,
@@ -276,14 +286,12 @@ class _RadioBtnElementWidgetState extends State<_RadioBtnElementWidget> {
         hoverColor: widget.elParams.p.disableParams!.hoverColor,
         highlightColor: widget.elParams.p.disableParams!.highlightColor,
         splashColor: widget.elParams.p.disableParams!.splashColor,
-        disabledColor: widget.elParams.p.enableTapLabel
-            ? widget.elParams.p.enableParams!.color ?? Colors.black87
-            : widget.elParams.p.disableParams!.disabledColor,
-        onPressed: widget.elParams.p.enableTapLabel
-            ? null
-            : () {
+        disabledColor: widget.elParams.p.disableParams!.disabledColor,
+        onPressed: widget.elParams.p.isEnabled
+            ? () {
                 _onTapCallback(index);
-              },
+              }
+            : null,
         mouseCursor: widget.elParams.p.disableParams!.mouseCursor,
         focusNode: widget.elParams.p.disableParams!.focusNode,
         autofocus: widget.elParams.p.disableParams!.autofocus,
@@ -308,14 +316,12 @@ class _RadioBtnElementWidgetState extends State<_RadioBtnElementWidget> {
         hoverColor: widget.elParams.p.disableParams!.hoverColor,
         highlightColor: widget.elParams.p.disableParams!.highlightColor,
         splashColor: widget.elParams.p.disableParams!.splashColor,
-        disabledColor: widget.elParams.p.enableTapLabel
-            ? widget.elParams.p.enableParams!.color ?? Colors.black87
-            : widget.elParams.p.disableParams!.disabledColor,
-        onPressed: widget.elParams.p.enableTapLabel
-            ? null
-            : () {
+        disabledColor: widget.elParams.p.disableParams!.disabledColor,
+        onPressed: widget.elParams.p.isEnabled
+            ? () {
                 _onTapCallback(index);
-              },
+              }
+            : null,
         mouseCursor: widget.elParams.p.disableParams!.mouseCursor,
         focusNode: widget.elParams.p.disableParams!.focusNode,
         autofocus: widget.elParams.p.disableParams!.autofocus,