+++ lib/src/element/sub/text/super_and_subscript_element.dart
@@ -189,8 +189,8 @@ class SuperAndSubscriptElement extends TextElement {
           locale: textParams.p.locale,
           softWrap: textParams.p.softWrap,
           overflow: textParams.p.overflow,
-          textScaleFactor: textParams.p.textScaleFactor ??
-              MediaQuery.of(context).textScaleFactor,
+          textScaler:
+              textParams.p.textScaler ?? MediaQuery.of(context).textScaler,
           maxLines: textParams.p.maxLines,
           semanticsLabel: textParams.p.semanticsLabel,
           textWidthBasis: textParams.p.textWidthBasis,
@@ -212,8 +212,8 @@ class SuperAndSubscriptElement extends TextElement {
           locale: textParams.p.locale,
           softWrap: textParams.p.softWrap,
           overflow: textParams.p.overflow,
-          textScaleFactor: textParams.p.textScaleFactor ??
-              MediaQuery.of(context).textScaleFactor,
+          textScaler:
+              textParams.p.textScaler ?? MediaQuery.of(context).textScaler,
           maxLines: textParams.p.maxLines,
           semanticsLabel: textParams.p.semanticsLabel,
           textWidthBasis: textParams.p.textWidthBasis,