+++ lib/src/element_params/sub/button/dropdown_btn_params.dart
@@ -52,4 +52,5 @@ class DropdownBtnParams {
   // Pre version, This is selectedIndex.
   IndexManager? manager;
   void Function(int index)? menuCallback;
+  bool isEnabled = true;
 }