+++ lib/src/element_params/sub/button/switch_btn_params.dart
@@ -34,4 +34,5 @@ class SwitchBtnParams {
   double? splashRadius;
   FocusNode? focusNode;
   bool autofocus = false;
+  bool isEnabled = true;
 }