+++ lib/src/element_params/sub/text/ruby_text_params.dart
@@ -25,7 +25,7 @@ class RubyTextParams {
   Locale? locale;
   bool? softWrap;
   TextOverflow? overflow;
-  double? textScaleFactor;
+  TextScaler? textScaler;
   int? maxLines;
   String? semanticsLabel;
   TextWidthBasis? textWidthBasis;